+++ src/main/java/kopo/poly/kpaas/dto/ContractDTO.java
@@ -11,9 +11,9 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ContractDTO {
 
-    private Long contractId;
-    private Long userId;
-    private Long countryId;
+    private Integer contractId;
+    private Integer userId;
+    private Integer countryId;
     private String originalFileUrl;
     private String ocrText;
     private String createdAt;
